+++ backend/src/models/report.rs
@@ -1,8 +1,8 @@
 // src/models/report.rs
 
 use serde::{Deserialize, Serialize};
-// 1. Reemplazamos sqlx::FromRow por el equivalente nativo de SeaORM
 use sea_orm::FromQueryResult;
+use sea_orm::prelude::Decimal;
 
 #[derive(Debug, Serialize, Deserialize, FromQueryResult)] 
 pub struct SaleLine {
@@ -12,8 +12,8 @@ pub struct SaleLine {
     pub employee_name: String,
     pub product_name: String,
     pub amount: i32,
-    pub sale_price: f64,
-    pub line_total: f64,
+    pub sale_price: Decimal,
+    pub line_total: Decimal,
 }
 
 #[derive(Debug, Serialize, Deserialize, FromQueryResult)]
@@ -21,7 +21,7 @@ pub struct SupplierProductCount {
     pub id_supplier: i32,
     pub supplier_name: String,
     pub products_count: i64,
-    pub avg_unit_price: Option<f64>,
+    pub avg_unit_price: Option<Decimal>,
 }
 
 #[derive(Debug, Serialize, Deserialize, FromQueryResult)] 
@@ -30,13 +30,13 @@ pub struct CategorySales {
     pub category_name: String,
     pub items_sold: i64,
     pub sales_count: i64,
-    pub total_revenue: f64,
+    pub total_revenue: Decimal,
 }
 
 #[derive(Debug, Serialize, Deserialize, FromQueryResult)] 
 pub struct TopClient {
     pub id_client: i32,
     pub client_name: String,
-    pub total_spent: f64,
+    pub total_spent: Decimal,
     pub sales_count: i64,
 }
